AmiTCP_NG 4.1.5
Roadshow compatibility, a stack that can explain itself, the fix for a network
that died until you rebooted, a correctness sweep of the whole codebase, and an
ARP that stops shouting.

Asynchronous socket events. AmiTCP 3.0b2 predates the AmiTCP V4 event API, so
this fork never had it: SO_EVENTMASK, SBTC_SIGEVENTMASK (tag 4 -- the header's
numbering jumped 3 to 6), the FD_* set, and a GetSocketEvents() that used to
always answer "nothing". A Roadshow-model client fails at its first
setsockopt, which is why Amiga Explorer could not connect; it works now,
confirmed on real hardware. 15 further missing SocketBaseTags land with it:
UDP checksums, IP forwarding and TTL, four ICMP controls, byte counters, log
and error hooks, address-change notification.

"The network dies until I reboot". Recovery depended on the device the stack
had lost contact with, and reads and writes shared a buffer pool only that
recovery refilled; once drained nothing was posted to the card, which is why no
counter ever moved. The stack now wakes itself on pool exhaustion, from the
watchdog, and at teardown. What made it permanent is fixed; what stalls it is
still unknown.

ARP pacing and reachability. The 4.3BSD base broadcast a fresh request for
EVERY packet sent to an unresolved address -- no rate limit, no cap -- and
returned success, so a destination that never answered produced an unbounded
broadcast storm and the application was told its data had been sent. Now: one
request a second, a bounded burst, then a hold-down that reports the host
unreachable, while still probing slowly so recovery needs no help from the
caller. A resolved entry is re-checked rather than trusted for its full twenty
minutes, by UNICAST probe to the address already held -- broadcast is what an
access point drops first, and revalidation that depends on it fails exactly
when it is needed. Ordinary inbound traffic confirms a peer is reachable,
matched on the frame's sender hardware address because that is the next hop an
ARP entry actually maps, so a connection carrying data is never probed at all.
An entry is only given up on after both the unicast and broadcast phases have
gone unanswered.

Logging, which had never written a byte in any version -- messages were
composed, queued and dropped. Timestamps called timer.device before it was
opened, a null jump that on a machine with no MMU executes whatever sits at
address 0. Three switches now: LOGGING (on), LOGLEVEL (5), LOGCONSOLE (off),
so a quiet file can be kept without a console window appearing. LOGLEVEL=7
also names any library call that fails and its errno; that ships in every
build, with no build flag, because a diagnostic that needs a special build
cannot diagnose a machine you cannot reach.

Correctness and memory safety, from a sweep of kern, netinet, net and api.
udp_input() never rejected a sender-supplied uh_ulen smaller than the UDP
header, so a forged length drove the signed m_len negative and corrupted the
socket's byte accounting -- a remote denial of service against any bound port,
needing no handshake and no valid checksum. The access-control table was
bounded against a compile-time constant after being shrink-wrapped to an exact
fit, so the first ARexx ADD ACCESS wrote past it with caller-supplied bytes,
and a netdb reload transplanted the table without its count. dTable and
dTableSize are a pair whose bitmap lives inside the same allocation, yet were
published unlocked with the old table freed first; errnoSize and errnoPtr had
the same shape with readers that are lock-free by design. rexx_poll() took
messages off the ARexx port and, when it could not parse them, dropped them
without replying -- stranding the sender forever.

Earlier in this revision: the first library call of any kind now starts the
stack (39 vectors did not, and the machine stopped silently); uploads got
their fast path back, 9% to 57% measured on hardware; one library base may be
shared between tasks; netstat, nslookup, ftp, tftp and sntp ship with the
stack; ICMP address-mask replies are off by default, per RFC 1122; and the
installer delivers new settings to existing installations, which it had been
silently failing to do.

Verified on an emulated 68000: 67 checks across two guest tests, plus a
negative control confirming the ARexx test fails when its fix is reverted. The
descriptor-table race has no reliable test -- a negative control passed with
the fix removed -- and the test records that rather than implying coverage.
The ARP hold-down and give-up paths need a peer that stops answering and so
are review-verified, not test-verified; the revalidation, unicast probing and
inbound-confirmation paths run continuously on real hardware.

+++ src/kern/amiga_subr.h
@@ -106,6 +106,17 @@ imax(int a, int b)
   return (a > b ? a : b);
 }
 
+/*
+ * WARNING: min()/max() take UNSIGNED parameters, unlike stock BSD where they
+ * are type-preserving macros. A negative argument converts to a huge unsigned
+ * value, so `max(x, LOWER_BOUND)` is a SILENT NO-OP when x is negative, and
+ * `min(x, y)` returns y for a negative x. Use imin()/imax() (or lmin()/lmax())
+ * whenever an operand can be negative -- a signed sequence-number difference, a
+ * length derived from a signed 16-bit protocol field, an sbspace() result, and
+ * so on. Two real bugs came from this (tcp_input.c's rcv_wnd and tcp_mss's
+ * sanity floor); the signatures are left as-is because most call sites are
+ * genuinely unsigned and changing them would be the riskier edit.
+ */
 static inline unsigned int
 min(unsigned int a, unsigned int b)
 {
@@ -237,7 +248,12 @@ strncpy(register char *s1, register const char *s2, register unsigned int len)
 static inline void
 aligned_bzero_const(void *buf, long size) 
 {
-  short lcount;
+  /* PORT (AmiTCP_NG): long, not short. A 16-bit signed longword count wraps for
+   * sizes at or above ~128 KB, which mbuf clusters and socket buffers can reach.
+   * This whole family is inert in the current build (USE_ALIGNED_COPIES is only
+   * set by the legacy SAS/C Smakefile), so this is a latent trap being closed,
+   * not a live bug. */
+  long lcount;
   long *lbuf = (long *)buf;
   short *sbuf;
 
@@ -277,7 +293,12 @@ aligned_bzero_const(void *buf, long size)
 static inline void
 aligned_bzero(void *buf, long size) 
 {
-  short lcount;
+  /* PORT (AmiTCP_NG): long, not short. A 16-bit signed longword count wraps for
+   * sizes at or above ~128 KB, which mbuf clusters and socket buffers can reach.
+   * This whole family is inert in the current build (USE_ALIGNED_COPIES is only
+   * set by the legacy SAS/C Smakefile), so this is a latent trap being closed,
+   * not a live bug. */
+  long lcount;
   long *lbuf = (long *)buf;
   short *sbuf;
 
@@ -299,7 +320,12 @@ aligned_bzero(void *buf, long size)
 static inline void
 aligned_bcopy_const(const void *src, void *dst, long size) 
 {
-  short lcount;
+  /* PORT (AmiTCP_NG): long, not short. A 16-bit signed longword count wraps for
+   * sizes at or above ~128 KB, which mbuf clusters and socket buffers can reach.
+   * This whole family is inert in the current build (USE_ALIGNED_COPIES is only
+   * set by the legacy SAS/C Smakefile), so this is a latent trap being closed,
+   * not a live bug. */
+  long lcount;
   long *ldst = (long *)dst;
   short *sdst;
   long *lsrc = (long *)src;
@@ -342,7 +368,12 @@ aligned_bcopy_const(const void *src, void *dst, long size)
 static inline void
 aligned_bcopy(const void *src, void *dst, long size) 
 {
-  short lcount;
+  /* PORT (AmiTCP_NG): long, not short. A 16-bit signed longword count wraps for
+   * sizes at or above ~128 KB, which mbuf clusters and socket buffers can reach.
+   * This whole family is inert in the current build (USE_ALIGNED_COPIES is only
+   * set by the legacy SAS/C Smakefile), so this is a latent trap being closed,
+   * not a live bug. */
+  long lcount;
   long *ldst = (long *)dst;
   short *sdst;
   long *lsrc = (long *)src;
